+++ spec/dummy/db/migrate/20260716000000_add_access_token_to_access_grants.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddAccessTokenToAccessGrants < ActiveRecord::Migration[6.1]
+  def change
+    add_reference :oauth_access_grants, :access_token, index: false
+  end
+end